Moved modules in Components folder
