+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - Unreleased
 ### Added
 - Added backtesting feature
+- Created Components module
 
 ## [1.1.0] - 2019-09-01
 ### Changed
