+++ src/Components/MarketProvider.py
@@ -9,7 +9,7 @@ currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentfram
 parentdir = os.path.dirname(currentdir)
 sys.path.insert(0, parentdir)
 
-from .Market import Market
+from Interfaces.Market import Market
 
 
 class MarketSource(Enum):
