+++ test/test_weighted_avg_peak.py
@@ -13,9 +13,9 @@ sys.path.insert(0, "{}/src".format(parentdir))
 from Strategies.WeightedAvgPeak import WeightedAvgPeak
 from Utility.Utils import TradeDirection
 from Interfaces.Market import Market
-from Interfaces.Broker import Broker
-from Interfaces.IGInterface import IGInterface
-from Interfaces.AVInterface import AVInterface
+from Components.Broker import Broker
+from Components.IGInterface import IGInterface
+from Components.AVInterface import AVInterface
 from common.MockRequests import (
     ig_request_login,
     ig_request_set_account,
